ui/requirements.txt
streamlit>=1.40,<2
PyYAML>=6.0
